🔨 Address SC runtime review feedback
