+++ docs/qdmi/index.md
@@ -6,7 +6,7 @@ provides a standardized interface for describing and interacting with quantum
 devices. This part of MQT Core contains the implementation of QDMI's different
 components, such as a [QDMI driver](driver.md), a
 [QDMI device for Neutral Atom Systems](na_device.md), a
-[QDMI device for superconducting systems](sc_device.md), and a
+[QDMI device for Superconducting Systems](sc_device.md), and a
 [QDMI device for a Classical Quantum Circuit Simulator](ddsim_device).
 
 ```{toctree}
