+++ include/mqt-core/qdmi/devices/sc/Device.hpp
@@ -8,6 +8,10 @@
  * Licensed under the MIT License
  */
 
+/** @file Device.hpp
+ * @brief The MQT QDMI device implementation for superconducting devices.
+ */
+
 #pragma once
 
 #include "mqt_sc_qdmi/device.h"
