+++ test/qdmi/devices/na/test_device.cpp
@@ -9,6 +9,7 @@
  */
 
 #include "mqt_na_qdmi/device.h"
+#include "qdmi/TestUtils.hpp"
 #include "qdmi/devices/na/Configuration.hpp"
 
 #include <gmock/gmock-matchers.h>
@@ -20,12 +21,10 @@
 #include <array>
 #include <cstddef>
 #include <cstdint>
-#include <cstdlib>
 #include <cstring>
 #include <fstream>
 #include <functional>
 #include <future>
-#include <optional>
 #include <sstream>
 #include <stdexcept>
 #include <string>
@@ -63,55 +62,7 @@ struct PairHash {
   }
 };
 
-class ScopedEnvironmentVariable {
-public:
-  ScopedEnvironmentVariable(std::string name, const std::string& value)
-      : name_(std::move(name)) {
-    if (const auto* previous = std::getenv(name_.c_str());
-        previous != nullptr) {
-      previous_ = previous;
-    }
-    set(value);
-  }
-
-  ~ScopedEnvironmentVariable() {
-    if (previous_) {
-      static_cast<void>(setWithoutChecking(*previous_));
-    } else {
-#ifdef _WIN32
-      static_cast<void>(_putenv_s(name_.c_str(), ""));
-#else
-      // NOLINTNEXTLINE(misc-include-cleaner)
-      static_cast<void>(unsetenv(name_.c_str()));
-#endif
-    }
-  }
-
-  ScopedEnvironmentVariable(const ScopedEnvironmentVariable&) = delete;
-  ScopedEnvironmentVariable&
-  operator=(const ScopedEnvironmentVariable&) = delete;
-  ScopedEnvironmentVariable(ScopedEnvironmentVariable&&) = delete;
-  ScopedEnvironmentVariable& operator=(ScopedEnvironmentVariable&&) = delete;
-
-private:
-  void set(const std::string& value) const {
-    if (!setWithoutChecking(value)) {
-      throw std::runtime_error("Failed to set environment variable " + name_);
-    }
-  }
-
-  [[nodiscard]] bool setWithoutChecking(const std::string& value) const {
-#ifdef _WIN32
-    return _putenv_s(name_.c_str(), value.c_str()) == 0;
-#else
-    // NOLINTNEXTLINE(misc-include-cleaner)
-    return setenv(name_.c_str(), value.c_str(), 1) == 0;
-#endif
-  }
-
-  std::string name_;
-  std::optional<std::string> previous_;
-};
+using mqt::test::ScopedEnvironmentVariable;
 
 [[nodiscard]] std::string queryName(MQT_NA_QDMI_Device_Session session) {
   size_t size = 0;
