+++ test/qdmi/registry/CMakeLists.txt
@@ -10,7 +10,8 @@ set(TARGET_NAME mqt-core-qdmi-registry-test)
 
 if(TARGET MQT::CoreQDMIDriver)
   package_add_test(${TARGET_NAME} MQT::CoreQDMIDriver test_device_registry.cpp)
-  target_include_directories(${TARGET_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src/qdmi/driver)
+  target_include_directories(${TARGET_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src/qdmi/driver
+                                                    ${PROJECT_SOURCE_DIR}/test)
   target_compile_definitions(
     ${TARGET_NAME}
     PRIVATE
