+++ test/qdmi/registry/test_device_registry.cpp
@@ -9,20 +9,18 @@
  */
 
 #include "DeviceRegistry.hpp"
+#include "qdmi/TestUtils.hpp"
 #include "qdmi/driver/Driver.hpp"
 
 #include <gtest/gtest.h>
 
 #include <algorithm>
-#include <cstdlib>
 #include <filesystem>
 #include <fstream>
-#include <optional>
 #include <random>
 #include <stdexcept>
 #include <string>
 #include <string_view>
-#include <utility>
 #include <variant>
 #include <vector>
 
@@ -56,55 +54,7 @@ private:
   std::filesystem::path path_;
 };
 
-class ScopedEnvironmentVariable {
-public:
-  ScopedEnvironmentVariable(std::string name, const std::string& value)
-      : name_(std::move(name)) {
-    if (const auto* previous = std::getenv(name_.c_str());
-        previous != nullptr) {
-      previous_ = previous;
-    }
-    set(value);
-  }
-
-  ~ScopedEnvironmentVariable() {
-    if (previous_) {
-      static_cast<void>(setWithoutChecking(*previous_));
-    } else {
-#ifdef _WIN32
-      static_cast<void>(_putenv_s(name_.c_str(), ""));
-#else
-      // NOLINTNEXTLINE(misc-include-cleaner)
-      static_cast<void>(unsetenv(name_.c_str()));
-#endif
-    }
-  }
-
-  ScopedEnvironmentVariable(const ScopedEnvironmentVariable&) = delete;
-  ScopedEnvironmentVariable&
-  operator=(const ScopedEnvironmentVariable&) = delete;
-  ScopedEnvironmentVariable(ScopedEnvironmentVariable&&) = delete;
-  ScopedEnvironmentVariable& operator=(ScopedEnvironmentVariable&&) = delete;
-
-private:
-  void set(const std::string& value) const {
-    if (!setWithoutChecking(value)) {
-      throw std::runtime_error("Failed to set environment variable " + name_);
-    }
-  }
-
-  [[nodiscard]] bool setWithoutChecking(const std::string& value) const {
-#ifdef _WIN32
-    return _putenv_s(name_.c_str(), value.c_str()) == 0;
-#else
-    // NOLINTNEXTLINE(misc-include-cleaner)
-    return setenv(name_.c_str(), value.c_str(), 1) == 0;
-#endif
-  }
-
-  std::string name_;
-  std::optional<std::string> previous_;
-};
+using mqt::test::ScopedEnvironmentVariable;
 
 class ScopedCurrentPath {
 public:
